处理check 逻辑

+++ src/main/java/com/ora/blockchain/mybatis/entity/wallet/ERC20Sum.java
@@ -2,13 +2,15 @@ package com.ora.blockchain.mybatis.entity.wallet;
 
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Setter
 @Getter
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
 public class ERC20Sum {
-    private Long sumValue;
+    private BigInteger sumValue;
 
-    private Long gasUsed;
+    private BigInteger gasUsed;
 }
